Update README.md
I am creating program to ask user their name and age and tell them in which year they will turn 100
I originally had a simpler version which provided an incorrect year if the user had not currently had their birthday, so I have added an if/else statement to clarify this. However whatever the input for birthday_already the "if" statement always executes. I assume that I have made a mistake in the "if" statement somewhere, or am using input() incorrectly. Can anyone help?